Refactor folder structure for dockerization
Changes:
- Moved ingestion to src
- Fixed failing unit tests

TODO Items:
- Introduced a bug in docker container to web backend. It keeps failing. Will fix later.
