+++ src/app/requirements.txt
@@ -2,3 +2,7 @@ fastapi~=0.115.8
 uvicorn>=0.15.0
 pydantic~=2.10.6
 langchain~=0.3.18
+langchain-core
+langchain-community
+tiktoken
+openai
