+++ src/ingestion/chroma_ingestion.py
@@ -4,7 +4,9 @@ for loading and processing given data to vector database.
 """
 import logging
 
-from ingestion.ingestion import Ingestion
+from langchain.retrievers import SelfQueryRetriever
+
+from src.ingestion.ingestion import Ingestion
 from langchain_community.vectorstores import Chroma
 
 
@@ -17,3 +19,15 @@ class ChromaIngestion(Ingestion):
         vectorstore = Chroma.from_documents(text_splits, self.embeddings)
         logging.info("Chroma index is ready for retrieval!")
         return vectorstore.as_retriever()
+
+    def retrieve_documents(self, query: str):
+        """
+        Retrieve documents related to the given query.
+        """
+        vectorstore = Chroma(
+            embedding_function=self.embeddings,
+        )
+
+        # Initialize the retriever
+        retriever = SelfQueryRetriever(vectorstore=vectorstore)
+        return retriever.invoke(query)
